Move ETOS API communication to its own module
