+++ cli/src/etos_client/etos/__init__.py
@@ -14,3 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """ETOS client module."""
+from .etos import ETOS
